feat(generator): Refactoring the Skill  generator; bumping version to 0.15.0-SNAPSHOT
