+++ skills-generator/src/test/java/info/jab/pml/SkillReferenceGeneratorTest.java
@@ -25,9 +25,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayName("Skill Reference Generator Tests")
-class CursorRulesGeneratorTest {
+class SkillReferenceGeneratorTest {
 
-    private static final Logger logger = LoggerFactory.getLogger(CursorRulesGeneratorTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(SkillReferenceGeneratorTest.class);
 
     @Nested
     @DisplayName("Parameterized Generate Method Tests")
@@ -71,7 +71,7 @@ class CursorRulesGeneratorTest {
          * Each base name corresponds to both an XML file and expected MDC file.
          */
         private static Stream<String> provideXmlFileNames() {
-            return SystemPromptsInventory.baseNames();
+            return SkillReferences.baseNames();
         }
 
         @ParameterizedTest
@@ -476,11 +476,63 @@ class CursorRulesGeneratorTest {
     }
 
     @Nested
+    @DisplayName("XML XInclude Link Validation Tests")
+    class XmlIncludeLinkValidationTests {
+
+        private static Stream<String> provideXmlFilenames() {
+            return SkillReferences.xmlFilenames();
+        }
+
+        @ParameterizedTest
+        @MethodSource("provideXmlFilenames")
+        @DisplayName("Should have no broken xi:include hrefs")
+        void should_notHaveBrokenIncludes_when_xmlFileHasXiIncludes(String xmlFilename) throws Exception {
+            InputStream xmlStream = getClass().getClassLoader().getResourceAsStream(xmlFilename);
+            assertThat(xmlStream)
+                .withFailMessage("XML resource not found on classpath: %s", xmlFilename)
+                .isNotNull();
+
+            DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+            factory.setNamespaceAware(true);
+            factory.setXIncludeAware(false);
+
+            DocumentBuilder builder = factory.newDocumentBuilder();
+            Document document = builder.parse(xmlStream);
+
+            NodeList includes = document.getElementsByTagNameNS("http://www.w3.org/2001/XInclude", "include");
+
+            String baseDir = xmlFilename.contains("/")
+                ? xmlFilename.substring(0, xmlFilename.lastIndexOf('/') + 1)
+                : "";
+
+            List<String> brokenLinks = new ArrayList<>();
+
+            for (int i = 0; i < includes.getLength(); i++) {
+                Element include = (Element) includes.item(i);
+                String href = include.getAttribute("href");
+
+                if (href != null && !href.isEmpty()) {
+                    String resolvedPath = baseDir + href;
+                    try (InputStream ref = getClass().getClassLoader().getResourceAsStream(resolvedPath)) {
+                        if (ref == null) {
+                            brokenLinks.add(href + " (resolved as: " + resolvedPath + ")");
+                        }
+                    }
+                }
+            }
+
+            assertThat(brokenLinks)
+                .withFailMessage("XML file '%s' has broken xi:include links: %s", xmlFilename, brokenLinks)
+                .isEmpty();
+        }
+    }
+
+    @Nested
     @DisplayName("Version Consistency Tests")
     class VersionConsistencyTests {
 
         @ParameterizedTest
-        @MethodSource("info.jab.pml.SystemPromptsInventory#baseNames")
+        @MethodSource("info.jab.pml.SkillReferences#baseNames")
         @DisplayName("Should have metadata version matching project version from parent pom.xml")
         void should_haveMetadataVersionMatchingProjectVersion(String baseFileName) throws Exception {
             String expectedVersion = readProjectVersionFromParentPom();
